Update control.lua
Commented out lines 36 and 37, as those were causing a bug when placing entities in map view, on other surfaces than where the player character currently was.  

I theorise this was due to the 'fast replace' emulation.  As far as I am aware fast replace can only happen in default view (ie: not map view) and some aspect of the emulated fast replace was trying to happen on the surface the player character was currently on instead.